Cprowesl removes renegade input field in response to Issue #29
To fix Issue #29, I added an id to the ControlPanel.js file input variable.  I gave it a value of 'input_field'.  I then updated the InstructionMenu.js file by adding a check for the input_field to the constructor for the object.  Therefore, if the input_field is present when the IM is spawned, it will be removed.  There is now a console output which mentions when the input_field is removed.